pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "cce"
version = "0.2.0"
description = "Confidence-Consistency Evaluation for Time Series Anomaly Detection"
readme = "README.md"
license = "MIT"
authors = [
    {name = "EmorZz1G", email = "csemor@mail.scut.edu.cn"}
]
maintainers = [
    {name = "EmorZz1G", email = "csemor@mail.scut.edu.cn"}
]
keywords = [
    "time-series",
    "anomaly-detection",
    "evaluation",
    "metrics",
    "machine-learning",
    "confidence-consistency"
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "Intended Audience :: Science/Research",
    
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "Topic :: Scientific/Engineering :: Information Analysis",
    "Topic :: Software Development :: Libraries :: Python Modules",
]
requires-python = ">=3.8"
dependencies = [
    "torch>=1.8.0",
    "numpy>=1.19.0",
    "scipy>=1.7.0",
    "scikit-learn>=1.0.0",
    "pandas>=1.3.0",
    "matplotlib>=3.3.0",
    "seaborn>=0.11.0",
    "tqdm>=4.60.0",
    "pyyaml>=5.4.0",
    "deprecated>=1.2.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=6.0",
    "pytest-cov>=2.0",
    "black>=21.0",
    "flake8>=3.8",
    "mypy>=0.800",
    "pre-commit>=2.0",
]
docs = [
    "sphinx>=4.0",
    "sphinx-rtd-theme>=1.0",
    "myst-parser>=0.15",
]
jupyter = [
    "jupyter>=1.0.0",
    "ipython>=7.0.0",
    "plotly>=5.0.0",
    "bokeh>=2.4.0",
    "streamlit>=1.20.0",
]
all = [
    "cce[dev,docs,jupyter]",
]

[project.urls]
Homepage = "https://github.com/EmorZz1G/CCE"
Documentation = "https://github.com/EmorZz1G/CCE#readme"
Repository = "https://github.com/EmorZz1G/CCE.git"
Issues = "https://github.com/EmorZz1G/CCE/issues"
Changelog = "https://github.com/EmorZz1G/CCE/blob/main/CHANGELOG.md"

[project.scripts]
cce = "cce.cli:main"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]
include =["cce", "evaluation", "metrics", "utils", "data_utils"]  # 显式列出所有顶级模块
exclude = ["*.egg-info", ".cce", "scripts", "tests"]


[tool.setuptools.package-data]
cce = ["*.yaml", "*.yml", "*.json", "*.txt"]

[tool.black]
line-length = 88
target-version = ['py38']
include = '\.pyi?$'
extend-exclude = '''
/(
  # directories
  \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | build
  | dist
)/
'''

[tool.mypy]
python_version = "3.8"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
disallow_untyped_decorators = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
warn_unreachable = true
strict_equality = true

[tool.pytest.ini_options]
testpaths = ["../tests"]
python_files = ["test_*.py", "*_test.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = [
    "--strict-markers",
    "--strict-config",
    "--verbose",
    "--cov=cce",
    "--cov-report=term-missing",
    "--cov-report=html",
    "--cov-report=xml",
]
markers = [
    "slow: marks tests as slow (deselect with '-m \"not slow\"')",
    "integration: marks tests as integration tests",
    "unit: marks tests as unit tests",
]

[tool.coverage.run]
source = ["/src"]
omit = [
    "*/tests/*",
    "*/test_*",
    "*/__pycache__/*",
    "*/*",
    "*/dist/*",
]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "if self.debug:",
    "if settings.DEBUG",
    "raise AssertionError",
    "raise NotImplementedError",
    "if 0:",
    "if __name__ == .__main__.:",
    "class .*\\bProtocol\\):",
    "@(abc\\.)?abstractmethod",
]